-Reverts moving OpenIdConnectOptions to shared and instead uses a POCO to bring it's configuration across to the dashboard
