+++ src/Aspire.Hosting/PublicAPI.Unshipped.txt
@@ -1,16 +1,4 @@
 #nullable enable
-Aspire.Dashboard.Configuration.OpenIdConnectOptions
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.GetNameClaimTypes() -> string![]!
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.GetUsernameClaimTypes() -> string![]!
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.NameClaimType.get -> string!
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.NameClaimType.set -> void
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.OpenIdConnectOptions() -> void
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.RequiredClaimType.get -> string!
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.RequiredClaimType.set -> void
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.RequiredClaimValue.get -> string!
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.RequiredClaimValue.set -> void
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.UsernameClaimType.get -> string!
-Aspire.Dashboard.Configuration.OpenIdConnectOptions.UsernameClaimType.set -> void
 Aspire.Hosting.ApplicationModel.ContainerLifetime.Session = 0 -> Aspire.Hosting.ApplicationModel.ContainerLifetime
 Aspire.Hosting.ApplicationModel.CustomResourceSnapshot.HealthStatus.get -> Microsoft.Extensions.Diagnostics.HealthChecks.HealthStatus?
 Aspire.Hosting.ApplicationModel.DockerfileBuildAnnotation
